+++ store/criterion.go
@@ -1,8 +1,34 @@
+/*	MIT License
+
+	Copyright (c) 2019 Tim Shannon
+
+	Permission is hereby granted, free of charge, to any person obtaining a copy
+	of this software and associated documentation files (the "Software"), to deal
+	in the Software without restriction, including without limitation the rights
+	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+	copies of the Software, and to permit persons to whom the Software is
+	furnished to do so, subject to the following conditions:
+
+	The above copyright notice and this permission notice shall be included in all
+	copies or substantial portions of the Software.
+
+	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+	SOFTWARE.
+*/
+// Code has multiple changes compared to original, but corresponds proper mentioning.
 package store
 
 import (
 	"fmt"
+	"math/big"
 	"reflect"
+	"strings"
+	"time"
 )
 
 type Criterion struct {
@@ -26,26 +52,23 @@ func (c *Criterion) Eq(value interface{}) *Query {
 	return c.query
 }
 
-func (c *Criterion) compare(rowValue, criterionValue interface{}) (int, error) {
-	if rowValue == nil || criterionValue == nil {
-		if rowValue == criterionValue {
+func (c *Criterion) compare(testedValue, criterionValue interface{}) (int, error) {
+	if testedValue == nil || criterionValue == nil {
+		if testedValue == criterionValue {
 			return 0, nil
 		}
-		return 0, &ErrTypeMismatch{rowValue, criterionValue}
+		return 0, &errTypeMismatch{testedValue, criterionValue}
 	}
 
-	value := rowValue
-
-	for reflect.TypeOf(value).Kind() == reflect.Ptr {
-		value = reflect.ValueOf(value).Elem().Interface()
+	for reflect.TypeOf(testedValue).Kind() == reflect.Ptr {
+		testedValue = reflect.ValueOf(testedValue).Elem().Interface()
 	}
 
-	other := criterionValue
-	for reflect.TypeOf(other).Kind() == reflect.Ptr {
-		other = reflect.ValueOf(other).Elem().Interface()
+	for reflect.TypeOf(criterionValue).Kind() == reflect.Ptr {
+		criterionValue = reflect.ValueOf(criterionValue).Elem().Interface()
 	}
 
-	return compare(value, other)
+	return compare(testedValue, criterionValue)
 }
 
 func (c *Criterion) match(testValue interface{}) (bool, error) {
@@ -61,11 +84,284 @@ func (c *Criterion) match(testValue interface{}) (bool, error) {
 	}
 }
 
-type ErrTypeMismatch struct { // ToDo: unexport?
+func traverseFieldPath(value reflect.Value, fieldPath string) (reflect.Value, error) {
+	fields := strings.Split(fieldPath, ".")
+
+	current := value // ToDo: Can `current` be deleted?
+	for i := range fields {
+		if current.Kind() == reflect.Ptr {
+			current = current.Elem()
+		}
+		current = current.FieldByName(fields[i])
+
+		if !current.IsValid() {
+			return reflect.Value{}, fmt.Errorf("instance field %s doesn't exist in type %s", fieldPath, value)
+		}
+	}
+	return current, nil
+}
+
+type errTypeMismatch struct {
 	Value interface{}
 	Other interface{}
 }
 
-func (e *ErrTypeMismatch) Error() string {
+func (e *errTypeMismatch) Error() string {
 	return fmt.Sprintf("%v (%T) cannot be compared with %v (%T)", e.Value, e.Value, e.Other, e.Other)
+}
+
+//Comparer compares a type against the encoded value in the store. The result should be 0 if current==other,
+// -1 if current < other, and +1 if current > other.
+// If a field in a struct doesn't specify a comparer, then the default comparison is used (convert to string and compare)
+// this interface is already handled for standard Go Types as well as more complex ones such as those in time and big
+// an error is returned if the type cannot be compared
+// The concrete type will always be passedin, not a pointer
+type Comparer interface {
+	Compare(other interface{}) (int, error)
+}
+
+func compare(value, other interface{}) (int, error) {
+	switch t := value.(type) {
+	case time.Time:
+		tother, ok := other.(time.Time)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(time.Time).Equal(tother) {
+			return 0, nil
+		}
+
+		if value.(time.Time).Before(tother) {
+			return -1, nil
+		}
+		return 1, nil
+	case big.Float:
+		o, ok := other.(big.Float)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		v := value.(big.Float)
+
+		return v.Cmp(&o), nil
+	case big.Int:
+		o, ok := other.(big.Int)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		v := value.(big.Int)
+
+		return v.Cmp(&o), nil
+	case big.Rat:
+		o, ok := other.(big.Rat)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		v := value.(big.Rat)
+
+		return v.Cmp(&o), nil
+	case int:
+		tother, ok := other.(int)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(int) == tother {
+			return 0, nil
+		}
+
+		if value.(int) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case int8:
+		tother, ok := other.(int8)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(int8) == tother {
+			return 0, nil
+		}
+
+		if value.(int8) < tother {
+			return -1, nil
+		}
+		return 1, nil
+
+	case int16:
+		tother, ok := other.(int16)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(int16) == tother {
+			return 0, nil
+		}
+
+		if value.(int16) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case int32:
+		tother, ok := other.(int32)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(int32) == tother {
+			return 0, nil
+		}
+
+		if value.(int32) < tother {
+			return -1, nil
+		}
+		return 1, nil
+
+	case int64:
+		tother, ok := other.(int64)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(int64) == tother {
+			return 0, nil
+		}
+
+		if value.(int64) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case uint:
+		tother, ok := other.(uint)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(uint) == tother {
+			return 0, nil
+		}
+
+		if value.(uint) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case uint8:
+		tother, ok := other.(uint8)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(uint8) == tother {
+			return 0, nil
+		}
+
+		if value.(uint8) < tother {
+			return -1, nil
+		}
+		return 1, nil
+
+	case uint16:
+		tother, ok := other.(uint16)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(uint16) == tother {
+			return 0, nil
+		}
+
+		if value.(uint16) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case uint32:
+		tother, ok := other.(uint32)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(uint32) == tother {
+			return 0, nil
+		}
+
+		if value.(uint32) < tother {
+			return -1, nil
+		}
+		return 1, nil
+
+	case uint64:
+		tother, ok := other.(uint64)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(uint64) == tother {
+			return 0, nil
+		}
+
+		if value.(uint64) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case float32:
+		tother, ok := other.(float32)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(float32) == tother {
+			return 0, nil
+		}
+
+		if value.(float32) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case float64:
+		tother, ok := other.(float64)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(float64) == tother {
+			return 0, nil
+		}
+
+		if value.(float64) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case string:
+		tother, ok := other.(string)
+		if !ok {
+			return 0, &errTypeMismatch{t, other}
+		}
+
+		if value.(string) == tother {
+			return 0, nil
+		}
+
+		if value.(string) < tother {
+			return -1, nil
+		}
+		return 1, nil
+	case Comparer:
+		return value.(Comparer).Compare(other)
+	default:
+		valS := fmt.Sprintf("%s", value)
+		otherS := fmt.Sprintf("%s", other)
+		if valS == otherS {
+			return 0, nil
+		}
+
+		if valS < otherS {
+			return -1, nil
+		}
+
+		return 1, nil
+	}
 }
